torch_ref: flip default dtype F64 → F32
Matches the BuildConfig flip in commit c01cd56b on the idris-ml
side. Apples-to-apples F32-vs-F32 ratios in bench-compare; no more
"Idris ran at F32 while ref ran at F64" caveat in perf comparisons.

Changes:
- training/runner.py: `_DTYPE = torch.float32` module default;
  `_dtype_for_device()` returns F32 unconditionally (kept as a fn
  for API stability — the 90 call sites in models/ all read
  through get_dtype() and pick up the flip automatically).
- bench_ops.py: 14 dtype=torch.float64 sites flipped to F32
  (matvec / matmul / softmax / linear-fwd / dot etc).
- compare_ops.py: banner "Precision: float64" → "float32".
- correctness/test_seq_classify.py: SeqClassifyCNN().double() →
  .float().

Untouched (deliberately F64):
- np.float64 in gym env state pins (env-physics needs F64
  precision per reference-alignment.md; CartPole loses ~5%
  convergence at F32 because the env's internal Euler step then
  runs in F32).
- bitlinear.py / bitnet.py F64 fixtures (bit-level reference
  tensors, not affected by training dtype).
- torch_ref/correctness/test_bitlinear.py F64 reference values.

**Known regressions surfaced by the flip** (5 of 102 tests fail at
F32, all RL/supervised convergence-threshold work — expected
rebaselining):

- correctness/test_a2c.py::test_returns_improve + test_converges:
  A2C on CartPole at F32 with the existing seed lands outside the
  convergence threshold; F32 gradient noise floor differs from
  F64 enough that the seed-tuned trajectory changes.
- correctness/test_ppo.py::test_returns_improve + test_converges:
  same class on Acrobot.
- correctness/test_supervised.py::test_converges: lands at 0.303
  vs threshold 0.3 (1% margin); tight either way, F32 noise
  pushed it over.

Filing as expected rebaselining work, not a regression to be
fixed in this commit. Per-test resolution paths (file individually
as follow-ups if desired):
- Adjust threshold to reflect F32 noise (loosen 0.3 → 0.35 for
  Supervised).
- Bump training budget (more epochs, more rollouts) to compensate
  for F32 noise.
- Multi-seed pass-rate metric instead of single-seed assertion
  (matches the convention in feedback_paired_side_alignment).
- Override per-test to keep F64 if the F32 numerical floor
  fundamentally breaks the convergence pattern.

For bench-compare purposes (the user's goal), correctness-test
failures are orthogonal: the bench harness times execution, not
convergence. The flip unblocks apples-to-apples F32-vs-F32
ratios.

+++ packages/pytorch/torch_ref/bench_ops.py
@@ -21,8 +21,8 @@ def wall_ms() -> float:
 
 
 def bench_matmul(m: int, n: int, k: int, iters: int) -> None:
-    a = torch.randn(m, n, dtype=torch.float64)
-    b = torch.randn(n, k, dtype=torch.float64)
+    a = torch.randn(m, n, dtype=torch.float32)
+    b = torch.randn(n, k, dtype=torch.float32)
 
     # warmup
     for _ in range(10):
@@ -40,8 +40,8 @@ def bench_matmul(m: int, n: int, k: int, iters: int) -> None:
 
 
 def bench_matvec(m: int, n: int, iters: int) -> None:
-    mat = torch.randn(m, n, dtype=torch.float64)
-    vec = torch.randn(n, dtype=torch.float64)
+    mat = torch.randn(m, n, dtype=torch.float32)
+    vec = torch.randn(n, dtype=torch.float32)
 
     for _ in range(10):
         _ = mat @ vec
@@ -58,8 +58,8 @@ def bench_matvec(m: int, n: int, iters: int) -> None:
 
 
 def bench_elementwise(n: int, iters: int) -> None:
-    a = torch.randn(n, dtype=torch.float64)
-    b = torch.randn(n, dtype=torch.float64)
+    a = torch.randn(n, dtype=torch.float32)
+    b = torch.randn(n, dtype=torch.float32)
 
     for _ in range(10):
         c = a + b
@@ -78,7 +78,7 @@ def bench_elementwise(n: int, iters: int) -> None:
 
 
 def bench_softmax(n: int, iters: int) -> None:
-    a = torch.randn(n, dtype=torch.float64)
+    a = torch.randn(n, dtype=torch.float32)
 
     for _ in range(10):
         _ = F.softmax(a, dim=0)
@@ -96,9 +96,9 @@ def bench_softmax(n: int, iters: int) -> None:
 
 def bench_conv2d(in_c: int, out_c: int, h: int, w: int, kh: int, kw: int, iters: int) -> None:
     # PyTorch conv2d expects [N, C, H, W] input
-    x = torch.randn(1, in_c, h, w, dtype=torch.float64)
-    weight = torch.randn(out_c, in_c, kh, kw, dtype=torch.float64)
-    bias = torch.randn(out_c, dtype=torch.float64)
+    x = torch.randn(1, in_c, h, w, dtype=torch.float32)
+    weight = torch.randn(out_c, in_c, kh, kw, dtype=torch.float32)
+    bias = torch.randn(out_c, dtype=torch.float32)
 
     for _ in range(2):
         _ = F.conv2d(x, weight, bias)
@@ -115,13 +115,13 @@ def bench_conv2d(in_c: int, out_c: int, h: int, w: int, kh: int, kw: int, iters:
 
 
 def bench_train_step(input_dim: int, output_dim: int, iters: int) -> None:
-    w = torch.randn(output_dim, input_dim, dtype=torch.float64, requires_grad=True)
-    b = torch.zeros(output_dim, dtype=torch.float64, requires_grad=True)
+    w = torch.randn(output_dim, input_dim, dtype=torch.float32, requires_grad=True)
+    b = torch.zeros(output_dim, dtype=torch.float32, requires_grad=True)
     opt = torch.optim.SGD([w, b], lr=0.01)
 
     # warmup
     for _ in range(5):
-        x = torch.randn(input_dim, dtype=torch.float64)
+        x = torch.randn(input_dim, dtype=torch.float32)
         y = w @ x + b
         loss = y.sum()
         opt.zero_grad()
@@ -130,7 +130,7 @@ def bench_train_step(input_dim: int, output_dim: int, iters: int) -> None:
 
     t0 = wall_ms()
     for _ in range(iters):
-        x = torch.randn(input_dim, dtype=torch.float64)
+        x = torch.randn(input_dim, dtype=torch.float32)
         y = w @ x + b
         loss = y.sum()
         opt.zero_grad()
